backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.10.1
cryptography==42.0.5
werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==1.0.1
nltk==3.8.1
sentence-transformers==2.2.2
chromadb==0.4.22
celery==5.3.6
redis==5.0.1
